Move pylxd.connection import pylxd.deprecated.connection
