+++ pylxd/deprecated/api.py
@@ -15,7 +15,7 @@
 import warnings
 
 from pylxd.deprecated import certificate
-from pylxd import connection
+from pylxd.deprecated import connection
 from pylxd.deprecated import container
 from pylxd import hosts
 from pylxd import image
