+++ pylxd/deprecated/connection.py
