+++ pylxd/tests/test_container.py
@@ -18,7 +18,7 @@ import json
 import mock
 import tempfile
 
-from pylxd import connection
+from pylxd.deprecated import connection
 
 from pylxd.tests import annotated_data
 from pylxd.tests import fake_api
