+++ pylxd/tests/test_network.py
@@ -15,7 +15,7 @@
 from ddt import ddt
 import mock
 
-from pylxd import connection
+from pylxd.deprecated import connection
 
 from pylxd.tests import annotated_data
 from pylxd.tests import fake_api
