+++ pylxd/tests/test_operation.py
@@ -16,7 +16,7 @@ import datetime
 from ddt import ddt
 import mock
 
-from pylxd import connection
+from pylxd.deprecated import connection
 
 from pylxd.tests import annotated_data
 from pylxd.tests import fake_api
